+++ examples/estimators/compute_abs_probs.py
@@ -28,8 +28,8 @@ g = cr.tl.estimators.GPCCA(k)
 # We need to compute or set the terminal states. In detail guide for both of our estimators can be found here
 # :ref:`sphx_glr_auto_examples_estimators_compute_terminal_states_gpcca.py`.
 g.compute_schur(n_components=4)
-g.compute_metastable_states(cluster_key="clusters")
-g.set_terminal_states_from_metastable_states(["Alpha", "Beta", "Epsilon"])
+g.compute_macrostates(cluster_key="clusters")
+g.set_terminal_states_from_macrostates(["Alpha", "Beta", "Epsilon"])
 
 # %%
 # :meth:`cellrank.tl.estimators.BaseEstimator.compute_absorption_probabilities` easily scales to 100k+ cells,