+++ tests/test_plotting.py
@@ -12,6 +12,8 @@ from anndata import AnnData
 
 import numpy as np
 import pandas as pd
+from scipy.sparse import issparse
+from pandas.api.types import is_categorical_dtype
 
 import matplotlib.cm as cm
 from matplotlib.testing import setup
@@ -58,10 +60,10 @@ RAW_GENES = [
 cr.settings.figdir = FIGS
 scv.settings.figdir = str(FIGS)
 
-
 try:
     from importlib_metadata import version as get_version
 except ImportError:
+    # >=Python3.8
     from importlib.metadata import version as get_version
 
 scvelo_paga_skip = pytest.mark.skipif(
@@ -513,7 +515,7 @@ class TestClusterLineages:
             adata,
             model,
             GENES[:10],
-            "0",
+            "1",
             time_key="latent_time",
             dpi=DPI,
             save=fpath,
@@ -540,7 +542,7 @@ class TestClusterLineages:
             adata,
             model,
             RAW_GENES[:5],
-            "0",
+            "1",
             time_key="latent_time",
             dpi=DPI,
             save=fpath,
@@ -554,7 +556,7 @@ class TestClusterLineages:
             adata,
             model,
             GENES[:10],
-            "0",
+            "1",
             time_key="latent_time",
             norm=False,
             dpi=DPI,
@@ -568,13 +570,85 @@ class TestClusterLineages:
             adata,
             model,
             GENES[:10],
-            "0",
+            "1",
             time_key="latent_time",
             data_key="Ms",
             norm=False,
             dpi=DPI,
             save=fpath,
         )
+
+    @compare()
+    def test_cluster_lineage_random_state(self, adata: AnnData, fpath: str):
+        model = create_model(adata)
+        cr.pl.cluster_lineage(
+            adata,
+            model,
+            GENES[:10],
+            "1",
+            time_key="latent_time",
+            random_state=42,
+            dpi=DPI,
+            save=fpath,
+        )
+
+    @compare()
+    def test_cluster_lineage_leiden(self, adata: AnnData, fpath: str):
+        model = create_model(adata)
+        cr.pl.cluster_lineage(
+            adata,
+            model,
+            GENES[:10],
+            "1",
+            time_key="latent_time",
+            use_leiden=True,
+            dpi=DPI,
+            save=fpath,
+        )
+
+    def test_cluster_lineage_random_state_same_pca(self, adata_cflare: AnnData):
+        model = create_model(adata_cflare)
+        cr.pl.cluster_lineage(
+            adata_cflare,
+            model,
+            GENES[:10],
+            "1",
+            time_key="latent_time",
+            random_state=42,
+            key="foo",
+        )
+
+        cr.pl.cluster_lineage(
+            adata_cflare,
+            model,
+            GENES[:10],
+            "1",
+            time_key="latent_time",
+            random_state=42,
+            key="bar",
+        )
+
+        np.allclose(
+            adata_cflare.uns["foo"].obsm["X_pca"], adata_cflare.uns["bar"].obsm["X_pca"]
+        )
+
+    def test_cluster_lineage_writes(self, adata_cflare: AnnData):
+        model = create_model(adata_cflare)
+        cr.pl.cluster_lineage(adata_cflare, model, GENES[:10], "0", n_test_points=200)
+
+        assert isinstance(adata_cflare.uns["lineage_0_trend"], AnnData)
+        assert adata_cflare.uns["lineage_0_trend"].shape == (10, 200)
+        assert is_categorical_dtype(adata_cflare.uns["lineage_0_trend"].obs["clusters"])
+
+    def test_cluster_lineage_key(self, adata_cflare: AnnData):
+        model = create_model(adata_cflare)
+        cr.pl.cluster_lineage(
+            adata_cflare, model, GENES[:10], "0", n_test_points=200, key="foobar"
+        )
+
+        assert isinstance(adata_cflare.uns["foobar"], AnnData)
+        assert adata_cflare.uns["foobar"].shape == (10, 200)
+        assert is_categorical_dtype(adata_cflare.uns["foobar"].obs["clusters"])
 
 
 class TestHeatmap:
@@ -1720,35 +1794,35 @@ class TestGPCCA:
 
     @compare(kind="gpcca")
     def test_scvelo_gpcca_meta_states(self, mc: GPCCA, fpath: str):
-        mc.plot_metastable_states(dpi=DPI, save=fpath)
+        mc.plot_macrostates(dpi=DPI, save=fpath)
 
     @compare(kind="gpcca")
     def test_scvelo_gpcca_meta_states_lineages(self, mc: GPCCA, fpath: str):
-        mc.plot_metastable_states(lineages=["0"], dpi=DPI, save=fpath)
+        mc.plot_macrostates(lineages=["0"], dpi=DPI, save=fpath)
 
     @compare(kind="gpcca")
     def test_scvelo_gpcca_meta_states_discrete(self, mc: GPCCA, fpath: str):
-        mc.plot_metastable_states(discrete=True, dpi=DPI, save=fpath)
+        mc.plot_macrostates(discrete=True, dpi=DPI, save=fpath)
 
     @compare(kind="gpcca")
     def test_scvelo_gpcca_meta_states_cluster_key(self, mc: GPCCA, fpath: str):
-        mc.plot_metastable_states(cluster_key="clusters", dpi=DPI, save=fpath)
+        mc.plot_macrostates(cluster_key="clusters", dpi=DPI, save=fpath)
 
     @compare(kind="gpcca")
     def test_scvelo_gpcca_meta_states_no_same_plot(self, mc: GPCCA, fpath: str):
-        mc.plot_metastable_states(same_plot=False, dpi=DPI, save=fpath)
+        mc.plot_macrostates(same_plot=False, dpi=DPI, save=fpath)
 
     @compare(kind="gpcca")
     def test_scvelo_gpcca_meta_states_cmap(self, mc: GPCCA, fpath: str):
-        mc.plot_metastable_states(cmap=cm.inferno, same_plot=False, dpi=DPI, save=fpath)
+        mc.plot_macrostates(cmap=cm.inferno, same_plot=False, dpi=DPI, save=fpath)
 
     @compare(kind="gpcca")
     def test_scvelo_gpcca_meta_states_title(self, mc: GPCCA, fpath: str):
-        mc.plot_metastable_states(title="foobar", dpi=DPI, save=fpath)
+        mc.plot_macrostates(title="foobar", dpi=DPI, save=fpath)
 
     @compare(kind="gpcca")
     def test_scvelo_gpcca_meta_states_time(self, mc: GPCCA, fpath: str):
-        mc.plot_metastable_states(mode="time", dpi=DPI, save=fpath)
+        mc.plot_macrostates(mode="time", dpi=DPI, save=fpath)
 
     @compare(kind="gpcca")
     def test_scvelo_gpcca_final_states(self, mc: GPCCA, fpath: str):
@@ -1984,11 +2058,12 @@ class TestLineageDrivers:
 
 
 # TODO: more model tests
+# TODO: fwd lineage 0 seems to be corrupted (i.e. very short, modify to 1)
 class TestModel:
     @compare()
     def test_model_default(self, adata: AnnData, fpath: str):
         model = create_model(adata)
-        model.prepare(adata.var_names[0], "0")
+        model.prepare(adata.var_names[0], "1")
         model.fit()
         model.predict()
         model.confidence_interval()
@@ -2003,8 +2078,26 @@ class TestModel:
         model.confidence_interval()
         model.plot(save=fpath)
 
+    @compare()
+    def test_model_obs_data_key(self, adata: AnnData, fpath: str):
+        model = create_model(adata)
+        gene = adata.X[:, 0]
+        adata.obs["foo"] = gene.A if issparse(gene) else gene
+
+        model.prepare("foo", "1", data_key="obs")
+        model.fit()
+        model.predict()
+        model.confidence_interval()
+        model.plot(save=fpath)
+
 
 class TestComposition:
     @compare()
     def test_composition(self, adata: AnnData, fpath: str):
         cr.pl._utils.composition(adata, "clusters", dpi=DPI, save=fpath)
+
+    @compare()
+    def test_composition_kwargs_autopct(self, adata: AnnData, fpath: str):
+        cr.pl._utils.composition(
+            adata, "clusters", dpi=DPI, save=fpath, autopct="%1.0f%%"
+        )